Adds an optional carriage return that enables reading CRLF-ending files
This is apparently required to enable tests to pass on a native Windows
clone of the source code, including in CI. Prior to this change, one
could run unix2dos on the test files to replicate the failure
I experienced on Windows even when tests pass on Linux for the v0.1.7
tag and fail on Windows.

I don't think I've fixed a Windows specific bug in anything in a decade!
This is my gaming rig, not programming, haha.