Let secretsignore have the highest precedence (#93)
Currently, the secretsignore specified files are added with the lowest precedence, so if some other regular gitignore file happens to unignore some pattern (i.e. with !pattern), its not easy to have the ripsecrets ignore these files.

This PR changes the behavior so that the file patterns from the secretsignore have the highest precedence.